refactor(app): split ServicesDashboardPage into smaller sibling files
Pure file-move refactor with no behavior change. Breaks the 1,759-line
ServicesDashboardPage.tsx monolith into a ServicesDashboardPage/ directory of
smaller cohesive files (largest resulting file: 699 lines).

An index.ts barrel re-exports the original public API (default page,
EndpointLatencyChart, buildInFilterCondition) so all external consumers resolve
unchanged.
