+++ packages/app/src/ServicesDashboardPage/index.ts
@@ -1,0 +1,3 @@
+export { buildInFilterCondition } from './helpers';
+export { EndpointLatencyChart } from './HttpTab';
+export { default } from './ServicesDashboardPage';
